requirements.txt
nltk >= 3.2.2
numpy >= 1.12.1
pandas >= 0.19.2
scipy >= 0.19.0
scikit-learn >= 0.18.1
python-dotenv >= 0.6.4
coverage >= 4.3.4
